move config convertion into ConfigHelper
